finish tidy up
