+++ lib/constants.js
@@ -1,0 +1,42 @@
+module.exports.devices = [
+  "Air Compressors / Pumps",
+  "Bellows",
+  "Self-inflating bags",
+  "Gas mixing valves",
+  "Pressure Regulators",
+  "Flow Control valves",
+  "Solenoid valves",
+  "Pressure relief valves",
+  "Check valves / one-way valves",
+  "Industrial Automation components (Safety Relays, PLCs)",
+  "Power Supplies",
+  "Electric Motors, and motor controllers",
+  "Linear actuators and controllers",
+  "Tubing and fittings",
+  "Pressure Sensors and Indicators",
+  "Oxygen Sensors and Indicators",
+  "Flow Sensors and Indicators",
+  "Manometers",
+  "Heat and moisture exchanging filters (HMEFs)",
+  "Air Filter, HEPA Filters"
+];
+
+module.exports.expertise = [
+  "Design / specification",
+  "Rapid prototyping",
+  "Manufacturing (manual)",
+  "Manufacturing (automated)",
+  "Machine Shops/sheet metal/tool manufacture",
+  "Pneumatic part manufacturers/suppliers",
+  "Contract/Product Assembly",
+  "Certification/regulation/testing",
+  "Logistics",
+  "Medical Training"
+];
+
+module.exports.resources = [
+  "Suitable space",
+  "Equipment",
+  "Trained personnel",
+  "Other",
+];
